requirements.txt
djangorestframework==3.15.2
psycopg2-binary
django-cors-headers==4.3.0
Pillow
djangorestframework-simplejwt==5.5.1
python-dotenv